Move polyfill code, Thanks to @qwtel
