+++ README.md
@@ -10,7 +10,7 @@ Native JavaScript [navigation](https://github.com/WICG/navigation-api) implement
 
  ### Test Coverage
 
- ![Web Platform Tests 129/237](https://img.shields.io/badge/Web%20Platform%20Tests-129%2F237-brightgreen) ![89.43%25 lines covered](https://img.shields.io/badge/lines-89.43%25-brightgreen) ![89.43%25 statements covered](https://img.shields.io/badge/statements-89.43%25-brightgreen) ![79.93%25 functions covered](https://img.shields.io/badge/functions-79.93%25-yellow) ![78.86%25 branches covered](https://img.shields.io/badge/branches-78.86%25-yellow)
+ ![Web Platform Tests 129/237](https://img.shields.io/badge/Web%20Platform%20Tests-129%2F237-brightgreen) ![89.31%25 lines covered](https://img.shields.io/badge/lines-89.31%25-brightgreen) ![89.31%25 statements covered](https://img.shields.io/badge/statements-89.31%25-brightgreen) ![80.85%25 functions covered](https://img.shields.io/badge/functions-80.85%25-brightgreen) ![79.43%25 branches covered](https://img.shields.io/badge/branches-79.43%25-yellow)
 
 [//]: # (badges)
